+++ lang/de/local_musi.php
@@ -61,6 +61,9 @@ $string['newsletterunsubscribed:error'] = 'Bei der Newsletter-Abmeldung ist ein 
 $string['birthdateprofilefield'] = 'Profilfeld für das Geburtsdatum';
 $string['birthdateprofilefielddesc'] = 'Wählen Sie das benutzerdefinierte Nutzerprofilfeld, in dem das Geburtsdatum gespeichert wird.';
 $string['autoaddtosubstitutionspool'] = 'LehrerInnen automatisch in VertreterInnen-Pool ihrer Sportart eintragen?';
+$string['musicachebookingoptionsettings'] = 'Booking-Options-Caching in M:USI-Tabellen für bessere Performance aktivieren';
+$string['musicachebookingoptionsanswers'] = 'Booking-Answers-Caching in M:USI-Tabellen für bessere Performance aktivieren';
+$string['musicacheexpirationtimeinseconds'] = 'Cache-Ablaufzeit in Sekunden (Beispiel: 3600 bedeutet, dass der Cache nach einer Stunde neu generiert wird)';
 
 // Shortcodes.
 $string['shortcodelists'] = 'Shortcode-Listen';
@@ -159,7 +162,7 @@ $string['changebookinginstance'] = 'Standard-Semester-Instanz setzen';
 $string['editbotags'] = 'Tags verwalten';
 $string['createbotag'] = 'Neuen Tag anlegen...';
 $string['createbotag:helptext'] = '<p>
-<a data-toggle="collapse" href="#collapseTagsHelptext" role="button" aria-expanded="false" aria-controls="collapseTagsHelptext">
+<a data-toggle="collapse" data-bs-toggle="collapse" href="#collapseTagsHelptext" role="button" aria-expanded="false" aria-controls="collapseTagsHelptext">
   <i class="fa fa-question-circle" aria-hidden="true"></i><span>&nbsp;Hilfe: So können Sie Tags konfigurieren...</span>
 </a>
 </p>
@@ -239,7 +242,6 @@ $string['bookingsalt'] = 'Anzahl der freien und maximal verfügbaren Kursplätze
 $string['teacheralt'] = 'Leiter des Kurses';
 $string['imagealt'] = 'Titelbild des Kurses';
 
-
 // Transactions List.
 $string['status'] = 'Status';
 $string['openorder'] = 'Offen';
@@ -282,7 +284,6 @@ $string['scheduler:tasklistdescription'] = 'Tasks welche zu einer bestimmten Zei
                 <code>[{"config": "schedulerenable", "scope" : "local_musi", "time" : "27.02.2024 12:00",
                 "value" : 0, "text" : "Deaktiviere Taskausführung um 12:00 Uhr"}]</code><br><br>Einmal abgearbeitet, wird der Task aus der
                 Taskliste entfernt';
-
 
 // Sports division.
 $string['nosportsdivision'] = 'Keine Sparten auf dieser Website verfügbar';