+++ settings.php
@@ -286,6 +286,34 @@ if ($hassiteconfig) {
         );
 
         $settings->add(
+            new admin_setting_configcheckbox(
+                'local_musi/musicachebookingoptionsettings',
+                get_string('musicachebookingoptionsettings', 'local_musi'),
+                '',
+                0
+            )
+        );
+
+        $settings->add(
+            new admin_setting_configcheckbox(
+                'local_musi/musicachebookingoptionsanswers',
+                get_string('musicachebookingoptionsanswers', 'local_musi'),
+                '',
+                0
+            )
+        );
+
+        $settings->add(
+            new admin_setting_configtext(
+                'local_musi/musicacheexpirationtimeinseconds',
+                get_string('musicacheexpirationtimeinseconds', 'local_musi'),
+                '',
+                3600,
+                PARAM_INT
+            )
+        );
+
+        $settings->add(
             new admin_setting_configtext(
                 'local_musi/shortcodesarchivecmids',
                 get_string('shortcodesarchivecmids', 'local_musi'),